# user/docs/swagger.yaml
definitions:
  users.User:
    properties:
      _id:
        type: string
      age:
        type: integer
      email:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/users:
    get:
      consumes:
      - application/json
      description: GetAll a new user item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.User'
            type: array
        "400":
          description: Bad Request
          schema:
            type: error
      summary: GetAll a user
      tags:
      - users
  /api/users/{code}:
    put:
      consumes:
      - application/json
      description: Update a new user item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: error
      summary: Update a user
      tags:
      - users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a new user item
      parameters:
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: error
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: GetOne a new user item
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            type: error
      summary: GetOne a user
      tags:
      - users
  /health:
    get:
      consumes:
      - '*/*'
      description: get the Health status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Show the Health status of server.
      tags:
      - Health Status
  /register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new user item
      parameters:
      - description: User's Name
        in: formData
        name: name
        required: true
        type: string
      - description: User's Email
        in: formData
        name: email
        required: true
        type: string
      - description: User's age
        in: formData
        name: age
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Create a user
      tags:
      - users
swagger: "2.0"